Renamed yeoman-command-base.ts to sfdx-yeoman-command.ts
Also includes changes to falcon/project/clone and falcon/project/create to update import references.
